CCP-2200 Use engage package, rename more menu item, fix readme markdown
