+++ frontend/components/MoreMenuItem/index.jsx
@@ -1,7 +1,6 @@
 import React from 'react';
 import PropTypes from 'prop-types';
-import Link from '@shopgate/pwa-common/components/Link';
-import I18n from '@shopgate/pwa-common/components/I18n';
+import { Link, I18n } from '@shopgate/engage/components';
 import styles from './style';
 
 /**
