+++ frontend/components/MoreMenuItem/style.js
@@ -1,5 +1,5 @@
 import { css } from 'glamor';
-import { themeConfig } from '@shopgate/pwa-common/helpers/config';
+import { themeConfig } from '@shopgate/engage';
 
 const { colors } = themeConfig;
