+++ frontend/portals/MyPointsIos/index.jsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import isIOSTheme from '@shopgate-ps/pwa-extension-kit/env/helpers/isIOSTheme';
 import AccountItem from '../../components/AccountItem';
-import MoreMenuItem from '../../components/MoreMenueItem';
+import MoreMenuItem from '../../components/MoreMenuItem';
 import { showMyPointsLink } from '../../config';
 
 export default props => (
